refactor:使用原生SQL语句替代了gorm框架中的一些函数

+++ DTO/post.go
@@ -1,10 +1,10 @@
 package DTO
 
 type PostDetail struct {
-	PostID        uint   `json:"post_id" db:"post_id"`
+	PostID        int64  `json:"post_id" db:"post_id"`
 	Title         string `json:"title" db:"title"`
 	Content       string `json:"content" db:"content"`
-	AuthorId      uint   `json:"author_id" db:"author_id"`
+	AuthorId      int64  `json:"author_id" db:"author_id"`
 	Username      string `json:"author_name" db:"username"`
 	CommunityID   int64  `json:"community_id" db:"community_id"`
 	CommunityName string `json:"community_name" db:"community_name"`
